fix: lint register-contest-participant
