+++ shared-lib/src/functions/registerContestParticipant.ts
@@ -1,0 +1,23 @@
+import { z } from "zod";
+
+export const RegisterContestParticipantFunctionRequestValidator = z.object({
+  secret: z.string(),
+});
+
+export type RegisterContestParticipantFunctionRequest = z.infer<
+  typeof RegisterContestParticipantFunctionRequestValidator
+>;
+
+export type RegisterContestParticipantFunctionResponse =
+  | ({
+      status: string;
+      message: string;
+    } & {
+      status: "success";
+      contestParticipantId: string;
+    })
+  | {
+      status: "error";
+      message: string;
+      error: string;
+    };
