+++ shared-lib/src/index.ts
@@ -5,3 +5,4 @@ export * from "./signatures";
 export * from "./types";
 export * from "./types/";
 export { GetPrivateKeyFunctionResponse } from "./functions/getPrivateKey";
+export { RegisterContestParticipantFunctionResponse } from "./functions/registerContestParticipant";
