+++ shared-lib/src/types/functions.ts
@@ -7,17 +7,3 @@ export type SubmitRallyFunctionResponse =
       status: "error";
       message: string;
     };
-
-export type RegisterContestParticipantFunctionResponse =
-  | ({
-      status: string;
-      message: string;
-    } & {
-      status: "success";
-      contestParticipantId: string;
-    })
-  | {
-      status: "error";
-      message: string;
-      error: string;
-    };
